feat: moved helper functions and getChangedFacilityFieldsForUpdate() to utils/facilityUtils.ts and moved checks for changes to facilitiesStore.ts
